LayoutTests/js/dom/dom-attributes-on-mismatch-type-expected.txt

CONSOLE MESSAGE: Deprecated attempt to access property 'href' on a non-HTMLAnchorElement object.
CONSOLE MESSAGE: Deprecated attempt to access property 'href' on a non-HTMLAnchorElement object.
Make sure DOM attributes handle different types of |this| properly

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS objectWithDocumentAsPrototype.all threw exception TypeError: Type error.
See what happens if we put the anchor prototype on a div
PASS testDiv instanceof HTMLAnchorElement is true
PASS testDiv.href is undefined.
PASS testDiv.href='error' threw exception TypeError: Type error.
PASS testDiv instanceof HTMLAnchorElement is true
PASS testDiv.href is undefined.
PASS testDiv.href='error' threw exception TypeError: Type error.
PASS successfullyParsed is true

TEST COMPLETE
